+++ src/control/nhk2025b_swerve_calculator/include/nhk2025b_swerve_calculator/swerve_calculator.hpp
@@ -14,8 +14,8 @@ class swerve_calculator : public rclcpp::Node {
    private:
     void twist_callback (const geometry_msgs::msg::TwistStamped::SharedPtr msg);
 
-    float wheel_radius;
-    float wheel_position;
+    double wheel_radius;
+    double wheel_position;
 
     rclcpp::Subscription<geometry_msgs::msg::TwistStamped>::SharedPtr twist_sub_;
     rclcpp::Publisher<nhk2025b_msgs::msg::Swerve>::SharedPtr          swerve_pub_;
