+++ src/visualization/nhk2025b_visualize_swerve/include/nhk2025b_visualize_swerve/visualize_swerve.hpp
@@ -11,10 +11,10 @@ class visualize_swerve : public rclcpp::Node {
     visualize_swerve (const rclcpp::NodeOptions& options);
 
    private:
-    float wheel_position;
-    float wheel_radius;
-    float last_angle[4];
-    void  swerve_callback (const nhk2025b_msgs::msg::Swerve::SharedPtr msg);
+    double wheel_position;
+    double wheel_radius;
+    double last_angle[4];
+    void   swerve_callback (const nhk2025b_msgs::msg::Swerve::SharedPtr msg);
 
     rclcpp::Publisher<visualization_msgs::msg::MarkerArray>::SharedPtr marker_publisher_;
     rclcpp::Subscription<nhk2025b_msgs::msg::Swerve>::SharedPtr        swerve_subscriber_;
